[ANDROSDK-2001] move api logic to network layer
