+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipConstraint.java
@@ -32,9 +32,6 @@ import android.database.Cursor;
 
 import androidx.annotation.Nullable;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import com.gabrielittner.auto.value.cursor.ColumnAdapter;
 import com.google.auto.value.AutoValue;
 
@@ -46,16 +43,13 @@ import org.hisp.dhis.android.core.common.BaseObject;
 import org.hisp.dhis.android.core.common.ObjectWithUid;
 
 @AutoValue
-@JsonDeserialize(builder = AutoValue_RelationshipConstraint.Builder.class)
 public abstract class RelationshipConstraint extends BaseObject {
 
     @Nullable
-    @JsonIgnore()
     @ColumnAdapter(ObjectWithUidColumnAdapter.class)
     public abstract ObjectWithUid relationshipType();
 
     @Nullable
-    @JsonIgnore()
     @ColumnAdapter(RelationshipConstraintTypeColumnAdapter.class)
     public abstract RelationshipConstraintType constraintType();
 
@@ -90,7 +84,6 @@ public abstract class RelationshipConstraint extends BaseObject {
     public abstract Builder toBuilder();
 
     @AutoValue.Builder
-    @JsonPOJOBuilder(withPrefix = "")
     public abstract static class Builder extends BaseObject.Builder<Builder> {
 
         public abstract Builder relationshipType(ObjectWithUid relationshipType);
